feat: correct detrending and plotting
